backend/requirementst.txt
django
django-cors-headers
django-ninja
django-ninja-jwt[crypto]
dj-database-url
psycopg[binary]
pydantic[email]
python-decouple
rav